add splittermond as a supported system
splittermond doesn't use the default health.max and health.value, but
instead health.max and health.total.value

that's why it needs to be customed added or the module doesn't work at
all

splittermond actually uses a more complicated approach for calculating
health. It uses temporary health, stun damage, and regular health on top
of a wound system with official names. It's good enough like this in my
opinion or it would be a whole new module just for splittermond