pull only fontFamily from CONFIG.canvasTextStyle
After doing some testing, it seems my pr had some issues which caused `userTextStyle`'s settings to actually override the `canvasTextStyle`. This pr fixes things so that doesn't happen, as I'm now only taking the fontFamily, not the styleID (Surprise! I did not realize this was a thing). My apologies, and thanks again for this great module.